update to avalonia 11.1.3

+++ DownloaderUI/ViewModels/SettingsPageViewModel.cs
@@ -168,7 +168,7 @@ namespace DownloaderUI.ViewModels
                         {
                             if (value != null)
                             {
-                                Application.Current.Resources["ControlContentThemeFontSize"] = FontSize;
+                                Application.Current.Resources["ControlContentThemeFontSize"] = (double)FontSize;
                             }
                             await SaveAsync();
                         }
